fix: passando os metodos para os serviços corretos
